Restructure proptests: separate model from SUT, named invariants, subdirectory
Per PR review feedback, refactor the property-based test suite to make
its structure self-documenting and the model/SUT boundary explicit.

# Structure

Move all proptest-related code into a `proptests/` subdirectory so it is
clearly separated from production source files:

* `proptests.rs` — module root with an architectural overview and a
  file map. Only docs and `mod` declarations.
* `proptests/grammar.rs` — `Op`, `WakerAction`, sequence strategies.
* `proptests/manual_event.rs` — `ManualEvent` trait + impls.
* `proptests/model.rs` — `Model` and `SlotModel`: reference state with
  transition methods (`set_effective`, `reset`, `add_slot`,
  `mark_ready`, `mark_dropped`), independent of SUT mechanics.
* `proptests/waker_tracker.rs` — `WakerTracker` (live observation) and
  the raw waker vtable.
* `proptests/harness.rs` — `Harness` driver: per-operation methods,
  slot lifecycle helpers that keep SUT and model index-aligned, and
  named invariant checks.
* `proptests/local_manual.rs`, `proptests/manual.rs` — entry points
  (proptest! blocks + `regression_tests`).

The old `proptest_harness.rs`, `local_manual_proptest.rs`, and
`manual_proptest.rs` are deleted; `lib.rs` exposes a single
`#[cfg(test)] mod proptests;`.

# Model / SUT split

`Slot<F>` (SUT-side) now holds only `future` and `tracker`. The
lifecycle state (`Pending` / `MustBeReady` / `Ready` / `Dropped`) and
the wake-budget mirror live in `SlotModel`. Helpers (`add_slot`,
`take_future_and_mark_dropped`, etc.) keep the two index-aligned, and
`check_invariants` asserts the structural alignment.

# Named invariant functions

The monolithic `check_invariants` is replaced by three explicit checks:

* `check_invariant_no_lost_notifications` — for `MustBeReady` slots,
  `wake_count >= budget`. `event.set()` is synchronous, so a lower
  count is a lost wake.
* `check_invariant_wake_budget` — for all slots, `wake_count <=
  budget`. Catches duplicate wakes and stale post-terminal wakes.
* `check_invariant_no_stale_registration` — for `Ready` / `Dropped`
  slots, `wake_count == budget`. The model freezes the budget at the
  observed `wake_count` at the moment of terminal transition, so any
  subsequent wake immediately violates this with a precise message.

Per-op transition postconditions (e.g. "fresh future may not return
Ready when unset") stay inline with comments naming the invariant they
enforce.

# Terminology unified to "invariant"

The mixed "property" / "invariant" vocabulary is now consistent:
everything the harness checks is an *invariant*. "Property" is reserved
for the proptest framework concept (the macro, the testing technique).

# Coverage improvements from review critique

* `no_lost_notifications` is a lower-bound check so duplicate wakes
  surface as wake-budget violations, not misleading lost-wake errors.
* `DropFirstOther` / `DropLastOther` capture the running tracker as
  `Weak<WakerTracker>` and exclude it via `Rc::ptr_eq`, honoring the
  "other" in the action name even when the harness draws the same
  waker as both invoker and candidate.
* `run_sequence` ends with a `reset(); set();` smoke phase after
  cleanup so any latent stale registration on a terminal slot fires
  and trips the frozen-budget check.

All 151 events tests pass on Windows and WSL; `just package=events
validate-local` passes on both.
