+++ packages/events/src/proptests/local_manual.rs
@@ -1,40 +1,39 @@
 //! Property-based test entry point for [`LocalManualResetEvent`].
 //!
-//! See [`crate::proptest_harness`] for the shared model and
-//! invariants. This module only wires the harness up to the
-//! single-threaded primitive and contains hand-written regression
+//! See the [`proptests`](crate::proptests) module overview for the
+//! shared model and invariants. This file only wires the harness up to
+//! the single-threaded primitive and contains hand-written regression
 //! tests for known-tricky reentrancy scenarios.
 
 use proptest::prelude::*;
 
 use crate::LocalManualResetEvent;
-use crate::proptest_harness::{ops_strategy, run_sequence};
+use crate::proptests::grammar::ops_strategy;
+use crate::proptests::harness::run_sequence;
 
 proptest! {
     #![proptest_config(ProptestConfig {
         // Keep CI time bounded. The pilot can be scaled up via the
-        // `PROPTEST_CASES` environment variable for deeper bug
-        // hunts.
+        // `PROPTEST_CASES` environment variable for deeper bug hunts.
         cases: 256,
         // Disable filesystem regression persistence: we do not want
-        // every developer to accumulate `.proptest-regressions`
-        // files in their checkout. Shrunk reproductions are still
-        // printed on failure, which is the important signal.
+        // every developer to accumulate `.proptest-regressions` files
+        // in their checkout. Shrunk reproductions are still printed on
+        // failure, which is the important signal.
         failure_persistence: None,
         ..ProptestConfig::default()
     })]
 
     /// Combined feasibility test: random operation sequences must
-    /// satisfy all four properties (no lost notifications, wake
+    /// satisfy all four invariants (no lost notifications, wake
     /// budget, no stale registration, no panics). Asserts in the
-    /// harness implement properties 1–3; property 4 is implicit in
+    /// harness implement invariants 1–3; invariant 4 is implicit in
     /// the absence of panics or aborts.
     #[test]
-    // Proptest sequences run hundreds of cases per invocation;
-    // each case allocates and drops several futures. Under Miri
-    // this takes minutes per test and provides no additional
-    // value over the dedicated multithreaded tests in the
-    // `manual` / `auto` modules.
+    // Proptest sequences run hundreds of cases per invocation; each
+    // case allocates and drops several futures. Under Miri this takes
+    // minutes per test and provides no additional value over the
+    // dedicated multithreaded tests in the `manual` / `auto` modules.
     #[cfg_attr(miri, ignore)]
     fn proptest_local_manual_reset_event(ops in ops_strategy()) {
         run_sequence::<LocalManualResetEvent>(&ops);
@@ -42,13 +41,14 @@ proptest! {
 }
 
 #[cfg(test)]
-mod harness_self_tests {
-    //! Sanity checks that exercise the harness itself against
-    //! known-good operation sequences. These run under Miri so
-    //! the harness's `unsafe` waker plumbing is validated.
+mod regression_tests {
+    //! Hand-written sequences that exercise the harness against
+    //! known-good scenarios. These run under Miri so the harness's
+    //! `unsafe` waker plumbing is validated.
 
     use crate::LocalManualResetEvent;
-    use crate::proptest_harness::{Op, WakerAction, run_sequence};
+    use crate::proptests::grammar::{Op, WakerAction};
+    use crate::proptests::harness::run_sequence;
 
     #[test]
     fn empty_sequence_is_valid() {
@@ -71,9 +71,8 @@ mod harness_self_tests {
 
     #[test]
     fn reentrant_reset_does_not_skip_awaiters() {
-        // Mirror of the production test: A's waker resets the
-        // event, but B (registered before set) must still be
-        // notified.
+        // Mirror of the production test: A's waker resets the event,
+        // but B (registered before set) must still be notified.
         run_sequence::<LocalManualResetEvent>(&[
             Op::Register(WakerAction::Reset),
             Op::Register(WakerAction::None),
@@ -99,8 +98,8 @@ mod harness_self_tests {
     #[test]
     fn reentrant_reset_then_register_remains_pending() {
         // The fresh waiter belongs to the new generation; the outer
-        // set's drain skips it. After all polls, the fresh future
-        // must remain Pending.
+        // set's drain skips it. After all polls, the fresh future must
+        // remain Pending.
         run_sequence::<LocalManualResetEvent>(&[
             Op::Register(WakerAction::ResetThenRegister),
             Op::Set,
